moves GuiceJunit4Runner to commons test module
