+++ runtime/src/main/java/com/flipkart/flux/constant/RuntimeConstants.java
@@ -27,20 +27,11 @@ public class RuntimeConstants {
 
 
     /**
-     * Root for all the configs.
-     */
-    public static final String CONFIG_ROOT = "packaged";
-
-   /**
      * Root for dashboard webapp configs.
      */
     public static final String DASHBOARD_VIEW = "dashboard";
     public static final String FSM_VIEW = "fsmview";
 
-    /**
-     * Configuration yml resource path.
-     */
-    public static final String CONFIGURATION_YML = CONFIG_ROOT + "/configuration.yml";
     public static final String STATE_MACHINE_RESOURCE_RELATIVE_PATH = "/machines";
 
     /** default timeout for a task */ //todo: move it to configuration file
